Separate AFHDS2A and AFHDS3 UI settings.
